Move S3 uploader component to separate file
This is to help reduce clutter and separate some imports and a constant
that are only needed in one place.
